Simple readonly fullscreen client
This implements a single vertical slice of functionality: we can receive
messages from the server and display them.

To achieve this, I introduced a `Client` type that manages communication
with the server proper and a `UI` interface that displays content.

The `tui` package implements `UI`, but its design leaves much to be desired.
I feel that it currently contains interleaved responsibilities for both rendering
what is visible on screen and actually drawing to the screen. The render phase
should really be separate, as it will grow in complexity as it starts needing
to handle more state.

The `tui` package is also somewhat difficult to test. I will continue to ponder
what can be done about that, but I'm stumped for the time being.