Add failing tests for HistoryState
I decided to introduce a thing called a `HistoryState` that is responsible
for managing the actual contents of the history view that the user sees.
This is where we make all of the business-logic decisions about how things
ought to be rendered.

I haven't implemented it, but I've written tests reflecting that I expect
it to `Render` the view contents to a `io.Writer`. You interact with it
by informing it of new messages with `New` and by asking it to update the
UI by `Render`ing to it.

Eventually, I hope that it will track the size of the screen and will handle
all of the line-wrapping and colorization crap that will be needed, but let's
start small.