Tests for HistoryState render dimension awareness
`HistoryState` needs to know the dimensions of the render area in order
to correctly display messages. This adds a method to notify it of
current dimensions and also a failing test for whether it constrains
the number of rendered messages to (at maximum) the height of the render
area.

This is pretty naiive, as messages will certainly span multiple lines, but
you gotta start somewhere. Tests for intelligent line-wrapping and the like
will follow soon.