Fetch recent message history from server automatically
When you join a server, now we automatically query for the root
message and all recent messages. This causes the UI to populate
with some kind of content by default (even if it's just the root
message).

Thanks to the message history sorting from the last commit or two,
the messages also render in the proper order automatically, even
if they were received out-of-order.