Prepare for sending replies
I added a boolean to the TUI to track whether we're in "editing mode"
and was gearing up for adding the editor itself when I noticed that
the client always ignored my first keypress after opening it.

It took me a while to hunt down this problem. It was actually a simple
comparison error in `HistoryState.New()`, but the process of debugging
it showed me that I should really make HistoryState more concurrent-safe.
With that in mind, there is now a single goroutine responsible for making
changes to the actual state of the `HistoryState`. All methods that mutate
the state just send an anonymous function to that goroutine over a dedicated
channel. This ensures that the state modifications occur atomically.

All `HistoryState` methods that did not involve state modification have been
changed so they act on a `HistoryState` instead of a `*HistoryState`. This will
ensure that they are unable to actually modify it.