Simplify render logic
Previously, I'd written my own line-wrapping algorithm that took
unicode width into account. That's a lot of code, and the gocui
framework actually contains a softwrap capability anyway.

In the interest of a small and understandable codebase, I've
removed the custom rendering code. It is available in the previous
commit should anyone ever need it again.