Implement scrolling the message history
It certainly isn't the most intuitive possible design, but I've added the ability to scroll the viewport onto the message history via the "left/right" keys. This
leaves the "up/down" pair for actually selecting a message to reply to.

I think this isn't super intuitive, but it's actually difficult to make the
cursor perform scrolling itself right now. The problem is that only the
`HistoryState` knows exactly which lines in the buffer correspond to the actual
selected message. To make moving the cursor scroll would require either making
the `HistoryState` manage the viewport (doable, not sure that I want to), or
making the `HistoryState` return the line numbers containing the cursor to
the `TUI`. This second option seems better, but I'm going to defer working
on it until the UI gives a better sense of the message tree. It seems more
important for Arbor to be visibly a tree than for Arbor's scrolling to be
smoother.