Stable sort and lengthen history
I noticed that messages sent at nearly the same time would randomly
swap places in the message history because the sort algorithm reordered
them. I changed to a stable sort so that no longer happens. Additionally,
I've started rendering everything into the view each time instead of
shortening to only things visible. This is less efficient in the long run,
but it makes the client much more usable in the short term.