Implement ancestry highlighting
This commit highlights each ancestor of the selected message. This exposes
the chat tree in a highly visual, consumable way.

Because the history just grows without bound and is iterated in order to
achieve this, performance will suffer the more messages that the client
has seen. However, working slowly is better than unimplemented. We'll get
there...