Implement Persist/Load via JSON multicodec
I chose multicodec so that we could easily upgrade the storage format
later on without hating ourselves too much. It doesn't solve all problems,
but I hope that it will help with some of them. Perhaps I'm simply borrowing
trouble though.

It will continue to work for as long as we store history as a serialized, ordered
list of messages. As soon as that changes, something better will be needed.

Also, the documentation of Load() currently lies by saying it will combine data
from multiple sources. This is my next objective though.