Only query for what we do not have
This change was intended to simply prevent the client from asking for
messages that were already available in the history, but the process
of trying to do that led me to realize some deeper structural flaws
within the code.

I've unified the concept of the "Client" as being a type that provides
services that can compose messages and can look up messages in an archive.
Now you don't pass around instances of a client/composer/archive separately,
but rather you pass around a client that embeds those services. I hope that
this will improve clarity in the code, as a Client contains all services
relevant to a single server connection. If and when we ever add multi-server
features, it will hopefully be as simple as having the TUI juggle multiple
client intances.