Send a [quit] message when you close the client
Right now, I don't have a good way to block when you try to send a message.
Generally, that's the exact opposite of what you want a UI to do. However,
that means that I can't guarantee that the message is actually sent before
destroying the TUI and ending the process. There are better ways to do this,
but they involve more substantial change. One thing at a time.