Integrate multi-line editor
This commit brings together a lot of functionality and refactoring work.

Firstly, it begins the removal of view-specific logic from the TUI struct.
Each view should really have a Controller-like entity managing its internal
state. Editor begins that process, but there's a ways to go yet on cleaning
up the mess in TUI.

Secondly, this commit adds a new layout function that constrains the height of
the chat history by the height of the editor. I anticipate that this will get
complicated quickly in the future as we add more views, but right now it's
pleasantly simple.

Thirdly, this commit changes a default keybinding. I'm very sad about this, but
I feel that I have no choice. In order to get "add a new line to my message" to
work with Alt+Enter, we had to disable the ability to interpret the escape key
as an individual keypress. Because of this, breaking out of the editor is now
Ctrl+\. I'm open to suggestions if anyone has better ideas for this.

There's still work to be done on integrating the Editor. We lost the helpful
title messages that it used to have in this commit, and it is now capable of
managing the message that it's replying to as internal state instead of needing
to use whatever is under the cursor. I think I'll need to add these features
before merging.