Make Editor track reply-to message id
This commit makes the Editor itself responsible for tracking what message
it is currently composing a response to. This allows us to respond to the
root message without moving the cursor (among other things).

As such, I've changed the implementation of `r` so that your cursor doesn't
jarringly jump to the beginning of the history. Additionally, you can now
see the username of who you are replying to within the title of your editor.