Move most chat history logic into tui/history.go
This is the first step in making the logic for controlling the chat
history's view independent of the overall TUI logic. They've been
comingled for a long time, and that has gradually made the TUI
code harder to maintain. I'm hoping to cleanly separate them, but
it will take more work to completely finish it.